chore: move styles to .styles.ts files [react-drawer & react-toast]
