+++ change/@fluentui-react-drawer-d5f3923f-b3d7-4cbb-b309-9e093e473f1e.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "chore: move styles to .styles.ts files [react-drawer & react-toast]",
+  "packageName": "@fluentui/react-drawer",
+  "email": "olfedias@microsoft.com",
+  "dependentChangeType": "patch"
+}
