+++ packages/react-components/react-drawer/src/components/DrawerFooter/index.ts
@@ -2,4 +2,4 @@ export * from './DrawerFooter';
 export * from './DrawerFooter.types';
 export * from './renderDrawerFooter';
 export * from './useDrawerFooter';
-export * from './useDrawerFooterStyles';
+export * from './useDrawerFooterStyles.styles';
