+++ packages/react-components/react-drawer/src/components/DrawerHeader/index.ts
@@ -2,4 +2,4 @@ export * from './DrawerHeader';
 export * from './DrawerHeader.types';
 export * from './renderDrawerHeader';
 export * from './useDrawerHeader';
-export * from './useDrawerHeaderStyles';
+export * from './useDrawerHeaderStyles.styles';
