+++ packages/react-components/react-toast/src/components/Toast.tsx
@@ -5,75 +5,15 @@
 import * as React from 'react';
 import { Transition } from 'react-transition-group';
 import { useIsomorphicLayoutEffect } from '@fluentui/react-utilities';
-import { makeStyles, mergeClasses, shorthands } from '@griffel/react';
-import { tokens } from '@fluentui/react-theme';
+import { mergeClasses } from '@griffel/react';
 import { useToast, Toast as ToastProps } from '../state';
 import { Timer } from './Timer';
 import { Announce } from '../AriaLive';
-
-const useStyles = makeStyles({
-  toast: {
-    boxSizing: 'border-box',
-    marginTop: '16px',
-    minHeight: '44px',
-    ...shorthands.borderRadius(tokens.borderRadiusMedium),
-    '--fui-toast-height': '44px',
-  },
-  enter: {
-    animationDuration: '200ms, 400ms',
-    animationDelay: '0ms, 200ms',
-    animationName: [
-      {
-        from: {
-          maxHeight: 0,
-          opacity: 0,
-          marginTop: 0,
-        },
-        to: {
-          marginTop: '16px',
-          opacity: 0,
-          maxHeight: 'var(--fui-toast-height)',
-        },
-      },
-      {
-        from: {
-          opacity: 0,
-        },
-        to: {
-          opacity: 1,
-        },
-      },
-    ],
-  },
-
-  exit: {
-    animationDuration: '400ms, 200ms',
-    animationDelay: '0ms, 400ms',
-    animationName: [
-      {
-        from: {
-          opacity: 1,
-        },
-        to: {
-          opacity: 0,
-        },
-      },
-      {
-        from: {
-          opacity: 0,
-        },
-        to: {
-          opacity: 0,
-          marginTop: 0,
-          maxHeight: 0,
-        },
-      },
-    ],
-  },
-});
+import { useToastStyles } from './Toast.styles';
 
 export const Toast: React.FC<ToastProps & { visible: boolean; announce: Announce }> = props => {
-  const styles = useStyles();
+  const styles = useToastStyles();
+
   const { visible, children, close, remove, updateId, announce, ...toastOptions } = props;
   const { timeout, politeness } = toastOptions;
   const { play, running, toastRef } = useToast<HTMLDivElement>({ ...props, content: children });
