+++ packages/react-components/react-toast/src/components/Toaster.styles.ts
@@ -7,5 +7,6 @@ import { makeStyles } from '@griffel/react';
 export const useToasterStyles = makeStyles({
   container: {
     position: 'fixed',
+    width: '292px',
   },
 });
